Reorganized inspection codebase
