+++ src/components/inspections/Grid/InspectionCardGrid/filterInspection.ts
@@ -1,5 +1,10 @@
-import { InspectionFilterBy } from "../ts/const/inspection.constant";
-import { Inspection } from "../ts/types/db.interfaces";
+import { Inspection } from "../../../../../ts/types/db.interfaces";
+export enum InspectionFilterBy {
+  ROOM_NAME = "Room Name",
+  FLOOR_NUMBER = "Floor Number",
+  BUILDING_NAME = "Building Name",
+  MEMBER_NAME = "Member Name",
+}
 
 export const filterInspection = (
   inspections: Inspection[],
