+++ src/components/inspections/Grid/InspectionCardGrid/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./InspectionCardGrid";
