+++ src/components/inspections/Grid/InspectionDateSelector/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./InspectionDateSelector";
