+++ src/components/inspections/Grid/InspectionGrid/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./InspectionGrid";
