+++ src/components/inspections/Grid/UncompletedCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UncompletedCard";
