+++ src/components/inspections/Planner/CleanTypeRadioGroup/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CleanTypeRadioGroup";
