+++ src/components/inspections/Planner/CleanTypeRadioGroup/props.ts
@@ -1,0 +1,6 @@
+import { CleanType } from "@prisma/client";
+
+export interface CleanTypeRadioGroupProps {
+  selected: CleanType;
+  handleChange: (value: CleanType) => void;
+}
