+++ src/components/inspections/Planner/EllipsisTextChip/EllipsisTextChip.tsx
@@ -1,0 +1,22 @@
+import { montserrat } from "../../../../../pages/theme";
+import { EllipsisTextChipProps } from "./props";
+
+const EllipsisTextChip = (props: EllipsisTextChipProps) => {
+  const { children, width } = props;
+
+  return (
+    <div
+      style={{
+        whiteSpace: "nowrap",
+        overflow: "hidden",
+        textOverflow: "ellipsis",
+        width: width,
+        fontFamily: montserrat.style.fontFamily,
+      }}
+    >
+      {children}
+    </div>
+  );
+};
+
+export default EllipsisTextChip;
