+++ src/components/inspections/Planner/EllipsisTextChip/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./EllipsisTextChip";
