+++ src/components/inspections/Planner/EllipsisTextChip/props.ts
@@ -1,0 +1,4 @@
+export interface EllipsisTextChipProps {
+  children: React.ReactNode;
+  width: string;
+}
