+++ src/components/inspections/Planner/InspectionPlanner/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./InspectionPlanner";
