+++ src/components/inspections/Planner/TeamMemberMultiSelect/props.ts
@@ -1,0 +1,10 @@
+export interface TeamMemberOptionProps {
+  key: number;
+  name: string;
+}
+
+export interface TeamMemberMultiSelectProps {
+  options: TeamMemberOptionProps[];
+  selected: TeamMemberOptionProps[];
+  handleChange: (value: TeamMemberOptionProps[]) => void;
+}
