+++ src/components/inspections/Reports/InspectionMaker/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./InspectionMaker";
