+++ store.ts
@@ -1,10 +1,9 @@
 import { configureStore } from "@reduxjs/toolkit";
 import memberFirstNameSlice from "./slices/memberProfileSlice";
-import inspectionSlice from "./slices/inspectionSlice";
 import buildingSelectSlice from "./slices/buildingSelectSlice";
 import roomSelectSlice from "./slices/roomSelectSlice";
-import inspectionsFetchSlice from "./slices/inspectionsFetchSlice";
-import inspectionSelectionSlice from "./slices/inspectionSelectionSlice";
+import inspectionsFetchSlice from "./src/components/inspections/Grid/inspectionsFetchSlice";
+import inspectionSelectionSlice from "./src/components/inspections/inspectionSelectionSlice";
 import inspectionMakerSlice from "./src/components/inspections/Reports/InspectionMakerSlice";
 
 //Importing our reducers
@@ -12,7 +11,6 @@ export const store = configureStore({
   reducer: {
     memberProfile: memberFirstNameSlice,
     buildingSelect: buildingSelectSlice,
-    inspectionData: inspectionSlice,
     roomSelect: roomSelectSlice,
     inspectionsFetchData: inspectionsFetchSlice,
     inspectionSelectionData: inspectionSelectionSlice,
