--- ts/const/inspection.constant.ts
@@ -1,6 +1,0 @@
-export enum InspectionFilterBy {
-  ROOM_NAME = "Room Name",
-  FLOOR_NUMBER = "Floor Number",
-  BUILDING_NAME = "Building Name",
-  MEMBER_NAME = "Member Name",
-}
